+++ prototyper/CHANGELOG.md
@@ -11,5 +11,5 @@ All notable changes to this project will be documented in this file. See [conven
 ### Modified
 - Refine CSR group comments.
 - fix(prototyper): temporary PMU fix for possible S-mode DTB modification
-
+- fix(prototyper): validate DBCN console shared memory range
 ### Removed